wm-ecmabot: Replace `return` hack with a different hack
In order to make it work like the console and Node REPL, in
other words, to not need a return statment to return a value,
it needs to be run in the execution context that, from glancing
at the spec, is only exposed via eval().

Both the browser console (e.g. chromium devtools) and Node REPL,
appear to work (as of writing) by using native bindings to V8
in a way that can't be done from within JavaScript.